Image Optimized and Layout Refactor (#357)
* image performance optimization

* fixed auth layout and declared prop interface

* moved layout from root
