+++ src/components/Team/style.css
@@ -1,5 +1,5 @@
 .bg {
-  background: url('../../../public/assets/arrow_bg.png') top/cover;
+  background: url('../../../public/assets/arrow_bg.webp') top/cover;
   background-size: 100% 100%;
   margin-bottom: 50px;
 }
